+++ key_event_thread.py
@@ -1,0 +1,16 @@
+from threading import Thread
+from pynput.keyboard import Listener, Key
+
+
+class KeyEventThread(Thread):
+    def __init__(self, run_cmd):
+        super().__init__()
+        self.run_cmd = run_cmd
+        self.listener = Listener(on_press=self.on_press)
+        self.listener.start()
+
+    def on_press(self, key):
+        if key == Key.enter:
+            self.run_cmd()
+
+
